backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose
pydantic
aiofiles
apscheduler
python-multipart
neo4j
gensim
nltk
numpy
torch
pdfplumber
python-docx